upd: add more helper funcs

+++ src/frontend/scanner/scanner.h
@@ -9,8 +9,8 @@ typedef struct Scanner
     size_t start; // Where the current token began.
     size_t position;
     size_t length; // Length of the source code.
-    int column;
-    int line;
+    size_t column;
+    size_t line;
 } Scanner;
 
 Scanner *scanner_init(char *source);
